+++ cadhy/__init__.py
@@ -8,7 +8,7 @@ import sys
 bl_info = {
     "name": "CADHY",
     "author": "CADHY Team",
-    "version": (0, 3, 8),
+    "version": (0, 3, 9),
     "blender": (4, 1, 0),
     "location": "View3D > Sidebar > CADHY",
     "description": "Parametric modeling toolkit for hydraulic infrastructure and CFD domain generation",